Changed how slackstatus command is used to set status
Previously the command was slackstatus set <status-name>. I think
it might be more natural to use slackstatus <status-name> since that
is the default option someone would want to use. But I'm also not sure
that it will work well since it might conflict with subcommands in the
future and it might require logic to ensure that no statuses carry the
same name as any of the subcommands.