CLI now sets up directory, files, and default values if they don't exist
The CLI will check to see if the directory ~/.config/ssucli exists.
It then checks files statuses.yaml and tokenconfig to see if they
exist.

If statuses.yaml does not, it creates a new file and fills it with
some default values.

If tokenconfig does not exist or it does exist but does not have a
slack token value in it, it prompts the user to paste in their personal
slack token and then saves the value to tokenconfig